pyproject.toml
[project]
name = "bridge-case-rpa-coingecko"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0"
]
